Create System_Setup_Instructions
Some of the parts of this file may not apply to the current license of the official GitHub repository. In cases of conflict within this file the license listed for the official GitHub repository will be used. Edit this file as required to add in your own license terms, support, update or other policies.